+++ Assets/Views/RenderClientsDataForProduct.cs
@@ -16,7 +16,7 @@ public class RenderClientsDataForProduct : View
 
         List<int> xs = new List<int>();
 
-        for (var i = startDate; i < endDate; i+= C.PERIOD) // 7 - period length
+        for (var i = startDate; i < endDate; i+= (int)C.PERIOD) // 7 - period length
             xs.Add(i);
 
         //Debug.Log("amount of dots: " + xs.Count);
@@ -62,7 +62,7 @@ public class RenderClientsDataForProduct : View
         var insertBefore = necessaryMetrics - metricsCount;
         for (var i = 0; i < insertBefore; i++)
         {
-            metrics.Add(GetMetricMockup(0, start + i * C.PERIOD));
+            metrics.Add(GetMetricMockup(0, start + i * (int)C.PERIOD));
         }
 
         // add current value
